Organize the online registry stuff a bit
