+++ src/ReaderApp/Tasks.cs
@@ -1,38 +1,23 @@
-using System.Collections.Immutable;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
-using System.Net;
 using System.Text;
 using System.Text.Json;
-using System.Text.Json.Nodes;
-using System.Web;
-using AngleSharp;
-using AngleSharp.Dom;
-using AngleSharp.Html.Dom;
-using AngleSharp.Io;
 using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Spreadsheet;
-using DocumentFormat.OpenXml.Wordprocessing;
-using Microsoft.Extensions.Configuration;
 using QuestPDF.Fluent;
 using QuestPDF.Infrastructure;
 using ReaderApp.ExcelBuilder;
+using ReaderApp.OnlineRegistry;
 using ScheduleLib;
 using ScheduleLib.Builders;
 using ScheduleLib.Generation;
-using ScheduleLib.Parsing;
 using ScheduleLib.Parsing.CourseName;
 using ScheduleLib.Parsing.GroupParser;
 using Column = DocumentFormat.OpenXml.Spreadsheet.Column;
 using Columns = DocumentFormat.OpenXml.Spreadsheet.Columns;
 using Font = DocumentFormat.OpenXml.Spreadsheet.Font;
-using Group = ScheduleLib.Group;
 using HorizontalAlignmentValues = DocumentFormat.OpenXml.Spreadsheet.HorizontalAlignmentValues;
-using IDocument = AngleSharp.Dom.IDocument;
-using NotSupportedException = System.NotSupportedException;
-using Table = DocumentFormat.OpenXml.Spreadsheet.Table;
 using VerticalAlignmentValues = DocumentFormat.OpenXml.Spreadsheet.VerticalAlignmentValues;
 
 namespace ReaderApp;
@@ -57,6 +42,37 @@ public struct AllTeacherExcelParams
     public required ParityDisplayHandler ParityDisplay;
     public required TimeSlotDisplayHandler TimeSlotDisplay;
     public required Schedule Schedule;
+    public required LessonTimeConfig TimeConfig;
+}
+
+public struct ParseStudyWeekWordDocParams
+{
+    public required string InputPath;
+}
+
+public struct AddLessonsToOnlineRegistryParams()
+{
+    public required CancellationToken CancellationToken;
+    /// <summary>
+    /// Will be initialized from the default source if not provided.
+    /// </summary>
+    public Credentials? Credentials = null;
+    /// <summary>
+    /// Will be initialized to the default config if not provided.
+    /// </summary>
+    public JsonSerializerOptions? JsonOptions;
+    /// <summary>
+    /// Will be initialized to the default values if not provided.
+    /// </summary>
+    public NamesConfig? Names = null;
+
+    public required Session Session;
+    public required Schedule Schedule;
+    public required IRegistryErrorHandler ErrorHandler;
+    public required CourseNameUnifierModule CourseNameUnifier;
+    public required GroupParseContext GroupParseContext;
+    public required LookupModule LookupModule;
+    public required IAllScheduledDateProvider DateProvider;
     public required LessonTimeConfig TimeConfig;
 }
 
@@ -935,1221 +951,18 @@ public static class Tasks
         };
     }
 
-    public struct AddLessonsToOnlineRegistryParams()
-    {
-        public required CancellationToken CancellationToken;
-        /// <summary>
-        /// Will be initialized from the default source if not provided.
-        /// </summary>
-        public Credentials? Credentials = null;
-        /// <summary>
-        /// Will be initialized to the default config if not provided.
-        /// </summary>
-        public JsonSerializerOptions? JsonOptions;
-        /// <summary>
-        /// Will be initialized to the default values if not provided.
-        /// </summary>
-        public NamesConfig Names = default;
-
-        public required Session Session;
-        public required Schedule Schedule;
-        public required ILogger Logger;
-        public required CourseFinder CourseFinder;
-        public required GroupParseContext GroupParseContext;
-        public required LookupModule LookupModule;
-        public required IAllScheduledDateProvider DateProvider;
-        public required LessonTimeConfig TimeConfig;
-    }
-
-    public sealed class CourseFinder
-    {
-        public required CourseNameUnifierModule Impl { get; init; }
-        public required LookupModule LookupModule { get; init; }
-
-        public CourseId? Find(string name)
-        {
-            var ret = Impl.Find(new()
-            {
-                Lookup = LookupModule,
-                CourseName = name,
-                ParseOptions = new()
-                {
-                    IgnorePunctuation = true,
-                },
-            });
-            return ret;
-        }
-    }
-
-    public interface ILogger
-    {
-        void CourseNotFound(string courseName);
-        void GroupNotFound(string groupName);
-        void LessonWithoutName();
-
-        // May want to pull this out.
-        void CustomLessonType(ReadOnlySpan<char> ch);
-    }
-
-    public struct NamesConfigSource()
-    {
-        public string TokensFile = "tokens.json";
-        public string TokenCookieName = "ForDecanat";
-        public string RegistryBaseUrl = "http://crd.usm.md/studregistry/";
-        public string RegistryLoginPath = "Account/Login";
-        public string LessonsPath = "LessonAttendance";
-
-        public readonly NamesConfig Build()
-        {
-            var reg = new Uri(RegistryBaseUrl);
-            var login = new Uri(reg, RegistryLoginPath);
-            var lessons = new Uri(reg, LessonsPath);
-            return new()
-            {
-                TokensFile = TokensFile,
-                TokenCookieName = TokenCookieName,
-                LoginUrl = login,
-                LessonsUrl = lessons,
-                BaseUrl = reg,
-            };
-        }
-    }
-
-    public struct NamesConfig
-    {
-        public static readonly NamesConfig Default = new NamesConfigSource().Build();
-
-        public required string TokensFile;
-        public required string TokenCookieName;
-        public required Uri BaseUrl;
-        public required Uri LoginUrl;
-        public required Uri LessonsUrl;
-
-        public readonly bool IsInitialized => LoginUrl != null;
-    }
-
-    public static JsonSerializerOptions DefaultJsonOptions = new()
-    {
-        IndentSize = 4,
-        WriteIndented = true,
-    };
-
-    public static async Task AddLessonsToOnlineRegistry(AddLessonsToOnlineRegistryParams p)
-    {
-        p.Credentials ??= GetCredentials();
-        if (!p.Names.IsInitialized)
-        {
-            p.Names = NamesConfig.Default;
-        }
-
-        var cookieProvider = new MemoryCookieProvider();
-        var cookieContainer = cookieProvider.Container;
-
-        using var handler = new HttpClientHandler();
-        handler.CookieContainer = cookieContainer;
-        handler.UseCookies = true;
-        handler.AllowAutoRedirect = false;
-
-        using var httpClient = new HttpClient(handler);
-
-        var config = Configuration.Default;
-        config = config.WithDefaultLoader();
-        config = config.With<ICookieProvider>(_ => cookieProvider);
-
-        using var browsingContext = BrowsingContext.New(config);
-
-        await InitializeToken();
-
-        var courseLinks = await QueryCourseLinks();
-        foreach (var courseLink in courseLinks)
-        {
-            var groupsUrl = courseLink.Url;
-            var groups = await QueryGroupLinksOfCourse(groupsUrl);
-            foreach (var group in groups)
-            {
-                var lessons = MatchLessons(new()
-                {
-                    Lookup = p.LookupModule.LessonsByCourse,
-                    Schedule = p.Schedule,
-                    CourseId = courseLink.CourseId,
-                    GroupId = group.GroupId,
-                    SubGroup = group.SubGroup,
-                });
-
-                var existingLessonInstances = await QueryLessonInstances(group.Uri);
-                var orderedLessonInstances = existingLessonInstances
-                    .OrderBy(x => x.DateTime)
-                    .ToArray();
-                _ = orderedLessonInstances;
-
-                // Figure out the exact dates that the lessons will occur.
-                var times = GetDateTimesOfScheduledLessons(new()
-                {
-                    Lessons = lessons,
-                    Schedule = p.Schedule,
-                    DateProvider = p.DateProvider,
-                    TimeConfig = p.TimeConfig,
-                }).OrderBy(x => x.DateTime);
-            }
-        }
-
-        return;
-
-        async Task<IEnumerable<LessonInstanceLink>> QueryLessonInstances(Uri groupUri)
-        {
-            var doc = await GetHtml(groupUri);
-            return Ret();
-
-            IEnumerable<LessonInstanceLink> Ret()
-            {
-                const string rowPath = """main > div:nth-of-type(3) > table > tbody > tr""";
-                var rows = doc.QuerySelectorAll(rowPath).Skip(1);
-                foreach (var row in rows)
-                {
-                    var cells = row.Children;
-                    // NOTE: these are going to throw an invalid cast if anything is weird with the nodes.
-                    var first = ProcessFirst();
-                    var editUri = ProcessEdit();
-                    yield return new()
-                    {
-                        EditUri = editUri,
-                        DateTime = first.DateTime,
-                        LessonType = first.LessonType,
-                    };
-                    continue;
-
-                    (LessonType LessonType, DateTime DateTime) ProcessFirst()
-                    {
-                        var dateTimeAndTypeCell = (IHtmlTableDataCellElement) cells[0];
-                        var children = dateTimeAndTypeCell.ChildNodes;
-
-                        LessonType lessonType;
-                        {
-                            var typeNode = children[^1];
-                            var typeText = typeNode.TextContent;
-                            lessonType = ParseLessonType(typeText, p.Logger);
-                        }
-
-                        DateTime dateTime;
-                        {
-                            var anchor = children.OfType<IHtmlAnchorElement>().First();
-                            var dateTimeText = anchor.Text;
-                            var span = dateTimeText.AsSpan();
-                            const string format = "dd.MM.yyyy HH.mm";
-                            bool success = DateTime.TryParseExact(
-                                format: format,
-                                s: span,
-                                provider: null,
-                                style: DateTimeStyles.AssumeLocal,
-                                result: out dateTime);
-                            if (!success)
-                            {
-                                throw new NotSupportedException("The date time didn't parse properly");
-                            }
-                        }
-
-                        return (lessonType, dateTime);
-                    }
-                    Uri ProcessEdit()
-                    {
-                        var editCell = (IHtmlTableDataCellElement) cells[2];
-                        var editAnchor = (IHtmlAnchorElement) editCell.Children[0];
-                        var ret = new Uri(editAnchor.Href);
-                        return ret;
-                    }
-                }
-            }
-        }
-
-        async Task<IEnumerable<GroupLink>> QueryGroupLinksOfCourse(Uri courseUrl)
-        {
-            var doc = await GetHtml(courseUrl);
-            // here, the format is different
-            // DJ2302ru(II)
-            // DJ2301
-            // IA2401fr
-            return Ret();
-
-            IEnumerable<GroupLink> Ret()
-            {
-                const string path = """form[name="lesson"] > div.row:nth-of-type(2) > div.col:nth-of-type(1) > div.row > a:nth-of-type(1)""";
-                var anchors = doc.QuerySelectorAll(path);
-                foreach (var el in anchors)
-                {
-                    var anchor = (IHtmlAnchorElement) el;
-                    var url = anchor.Href;
-                    var groupName = anchor.Text;
-                    var groupForSearch = ParseGroupFromOnlineRegistry(p.GroupParseContext, groupName);
-                    var groupId = FindGroupMatch(p.Schedule, groupForSearch);
-                    if (groupId == GroupId.Invalid)
-                    {
-                        p.Logger.GroupNotFound(groupName);
-                        continue;
-                    }
-
-                    var uri = new Uri(url);
-                    var subgroup = new SubGroup(groupForSearch.SubGroupName.ToString());
-                    yield return new()
-                    {
-                        Uri = uri,
-                        GroupId = groupId,
-                        SubGroup = subgroup,
-                    };
-                }
-            }
-        }
-
-        async Task<IEnumerable<CourseLink>> QueryCourseLinks()
-        {
-            var doc = await GetHtml(p.Names.LessonsUrl);
-            var ret = Ret();
-            return ret;
-
-            IEnumerable<CourseLink> Ret()
-            {
-                var anchors = doc.QuerySelectorAll($"#nav-{SemString()} > div > span:nth-of-type(2) > a");
-                foreach (var el in anchors)
-                {
-                    var anchor = (IHtmlAnchorElement) el;
-                    var url = anchor.Href;
-                    var courseName = anchor.Text;
-                    if (courseName.Length == 0)
-                    {
-                        p.Logger.LessonWithoutName();
-                        continue;
-                    }
-                    if (p.CourseFinder.Find(courseName) is not { } courseId)
-                    {
-                        p.Logger.CourseNotFound(courseName);
-                        continue;
-                    }
-                    yield return new(courseId, new(url));
-                }
-            }
-
-            string SemString()
-            {
-                return p.Session switch
-                {
-                    Session.Ses1 => "1",
-                    Session.Ses2 => "2",
-                    _ => throw new InvalidOperationException("??"),
-                };
-            }
-        }
-
-        [SuppressMessage("ReSharper", "AccessToDisposedClosure")]
-        async Task<IDocument> GetHtml(Uri uri)
-        {
-            bool failedOnce = false;
-            while (true)
-            {
-                using var req = await httpClient.GetAsync(uri, cancellationToken: p.CancellationToken);
-                // if invalid token
-                if (req.StatusCode
-                    is HttpStatusCode.Unauthorized
-                    or HttpStatusCode.Redirect)
-                {
-                    if (failedOnce)
-                    {
-                        throw new InvalidOperationException("Failed to use the password to log in once.");
-                    }
-
-                    await QueryTokenAndSave();
-                    failedOnce = true;
-                    continue;
-                }
-                await using var res = await req.Content.ReadAsStreamAsync(p.CancellationToken);
-                var document = await browsingContext.OpenAsync(r => r.Content(res).Address(uri));
-                return document;
-            }
-        }
-
-        async Task InitializeToken()
-        {
-            if (await MaybeSetCookieFromFile())
-            {
-                return;
-            }
-            await QueryTokenAndSave();
-        }
-
-        async Task<bool> MaybeSetCookieFromFile()
-        {
-            var token = await LoadToken();
-            if (token is null)
-            {
-                return false;
-            }
-            if (token.Expired)
-            {
-                return false;
-            }
-            cookieContainer.Add(p.Names.BaseUrl, token);
-            return true;
-        }
-
-        async ValueTask<Cookie?> LoadToken()
-        {
-            if (!File.Exists(p.Names.TokensFile))
-            {
-                return null;
-            }
-
-            await using var stream = File.OpenRead(p.Names.TokensFile);
-            if (stream.Length == 0)
-            {
-                return null;
-            }
-
-            JsonDocument cookies;
-            try
-            {
-                cookies = await JsonDocument.ParseAsync(
-                    stream,
-                    cancellationToken: p.CancellationToken);
-                if (cookies == null)
-                {
-                    return null;
-                }
-            }
-            catch (JsonException)
-            {
-                stream.Close();
-                File.Delete(p.Names.TokensFile);
-                return null;
-            }
-
-            using var cookies_ = cookies;
-
-            var root = cookies.RootElement;
-            if (root.ValueKind != JsonValueKind.Object)
-            {
-                return null;
-            }
-            if (!root.TryGetProperty(p.Credentials.Login, out var token))
-            {
-                return null;
-            }
-
-            try
-            {
-                var cookie = token.Deserialize<TokenCookieModel>();
-                if (cookie is null)
-                {
-                    return null;
-                }
-                return cookie.ToObject(p.Names.TokenCookieName);
-            }
-            catch (JsonException)
-            {
-                return null;
-            }
-        }
-
-        [SuppressMessage("ReSharper", "AccessToDisposedClosure")]
-        async Task QueryTokenAndSave()
-        {
-            var success = await LogIn(httpClient);
-            if (!success)
-            {
-                throw new InvalidOperationException("Login failed.");
-            }
-
-            var cookies = cookieContainer.GetCookies(p.Names.LoginUrl);
-            if (cookies[p.Names.TokenCookieName] is not { } token)
-            {
-                throw new InvalidOperationException("Token cookie not found.");
-            }
-
-            await using var stream = File.Open(p.Names.TokensFile, FileMode.OpenOrCreate, FileAccess.ReadWrite);
-            if (await TryUpdateExisting())
-            {
-                return;
-            }
-            await CreateNew();
-            return;
-
-            async ValueTask<bool> TryUpdateExisting()
-            {
-                if (stream.Length == 0)
-                {
-                    return false;
-                }
-
-                JsonNode? document;
-                try
-                {
-                    document = await JsonNode.ParseAsync(
-                        stream,
-                        cancellationToken: p.CancellationToken);
-                }
-                catch (JsonException)
-                {
-                    return false;
-                }
-
-                if (document is not JsonObject jobj)
-                {
-                    return false;
-                }
-                return await Save(jobj);
-            }
-
-            async Task<bool> CreateNew()
-            {
-                var root = new JsonObject();
-                return await Save(root);
-            }
-
-            async Task<bool> Save(JsonObject root)
-            {
-                stream.Seek(0, SeekOrigin.Begin);
-                var tokenModel = TokenCookieModel.FromObject(token);
-                root[p.Credentials.Login] = JsonSerializer.SerializeToNode(tokenModel);
-                await JsonSerializer.SerializeAsync(
-                    stream,
-                    root,
-                    options: p.JsonOptions ?? DefaultJsonOptions,
-                    cancellationToken: p.CancellationToken);
-                stream.SetLength(stream.Position);
-                return true;
-            }
-        }
-
-        async Task<bool> LogIn(HttpClient client)
-        {
-            Uri uri;
-            {
-                var b = new UriBuilder(p.Names.LoginUrl);
-                b.Port = -1;
-                var parameters = HttpUtility.ParseQueryString("");
-                parameters.Add("UserLogin", p.Credentials.Login);
-                parameters.Add("UserPassword", p.Credentials.Password);
-                b.Query = parameters.ToString();
-                uri = b.Uri;
-            }
-
-            var response = await client.PostAsync(
-                uri,
-                content: null,
-                cancellationToken: p.CancellationToken);
-            bool success = response.StatusCode == HttpStatusCode.Redirect;
-            return success;
-        }
-    }
-
-    public static Credentials GetCredentials()
-    {
-        var b = new ConfigurationBuilder();
-        b.AddUserSecrets<Program>();
-        var config = b.Build();
-        var ret = config.GetRequiredSection("Registry").Get<Credentials>();
-        if (ret == null)
-        {
-            throw new InvalidOperationException("Credentials not found.");
-        }
-        if (ret.Login == null)
-        {
-            throw new InvalidOperationException("Login not found.");
-        }
-        if (ret.Password == null)
-        {
-            throw new InvalidOperationException("Password not found.");
-        }
-        return ret;
-    }
-
-    static GroupForSearch ParseGroupFromOnlineRegistry(GroupParseContext context, string s)
-    {
-        var mainParser = new Parser(s);
-        mainParser.SkipWhitespace();
-
-        var bparser = mainParser.BufferedView();
-
-        var label = ParseLabel(ref bparser);
-        var year = ParseYear(ref bparser);
-        var groupNumber = ParseGroupNumber(ref bparser);
-        var nameWithoutFR = mainParser.PeekSpanUntilPosition(bparser.Position);
-        _ = nameWithoutFR;
-
-        mainParser.MoveTo(bparser.Position);
-
-        var languageOrFR = ParseLanguageOrFR(ref mainParser);
-        var subGroup = ParseSubGroup(ref mainParser);
-
-        mainParser.SkipWhitespace();
-        if (!mainParser.IsEmpty)
-        {
-            throw new NotSupportedException("Group name not parsed fully.");
-        }
-
-        var grade = context.DetermineGrade((int) year);
-
-        return new()
-        {
-            Grade = grade,
-            FacultyName = label,
-            GroupNumber = (int) groupNumber,
-            // Don't have precedents for master yet.
-            QualificationType = QualificationType.Licenta,
-            AttendanceMode = languageOrFR.FR ? AttendanceMode.FrecventaRedusa : AttendanceMode.Zi,
-            SubGroupName = subGroup,
-        };
-
-        static ReadOnlyMemory<char> ParseLabel(ref Parser parser)
-        {
-            if (!parser.CanPeekCount(2))
-            {
-                JustThrow("group label");
-            }
-            var ret = parser.PeekSource(2);
-            parser.Move(2);
-            return ret;
-        }
-
-        static uint ParseYear(ref Parser parser)
-        {
-            var yearResult = parser.ConsumePositiveInt(GroupHelper.YearLen);
-            if (yearResult.Status != ConsumeIntStatus.Ok)
-            {
-                JustThrow("year");
-            }
-            return yearResult.Value;
-        }
-
-        static uint ParseGroupNumber(ref Parser parser)
-        {
-            var numberResult = parser.ConsumePositiveInt(GroupHelper.GroupNumberLen);
-            if (numberResult.Status != ConsumeIntStatus.Ok)
-            {
-                JustThrow("group number");
-            }
-            return numberResult.Value;
-        }
-
-        static LanguageOrFR ParseLanguageOrFR(ref Parser parser)
-        {
-            var bparser = parser.BufferedView();
-            var skipResult = bparser.Skip(new SkipUntilOpenParenOrWhiteSpace());
-            if (!skipResult.SkippedAny)
-            {
-                return default;
-            }
-
-            var languageOrFRName = parser.PeekSpanUntilPosition(bparser.Position);
-            var ret = DetermineIfLabelIsLanguageOrFR(languageOrFRName);
-            parser.MoveTo(bparser.Position);
-            return ret;
-        }
-
-        static LanguageOrFR DetermineIfLabelIsLanguageOrFR(ReadOnlySpan<char> languageOrFRName)
-        {
-            LanguageOrFR ret = default;
-            if (languageOrFRName.Equals("fr", StringComparison.OrdinalIgnoreCase))
-            {
-                ret.FR = true;
-                return ret;
-            }
-
-            var maybeLang = LanguageHelper.ParseName(languageOrFRName);
-            if (maybeLang is not { } lang)
-            {
-                JustThrow("language");
-            }
-            ret.Language = lang;
-            return ret;
-        }
-
-        static ReadOnlyMemory<char> ParseSubGroup(ref Parser parser)
-        {
-            if (parser.IsEmpty)
-            {
-                return ReadOnlyMemory<char>.Empty;
-            }
-
-            // Possible if we're at a whitespace.
-            if (parser.Current != '(')
-            {
-                return ReadOnlyMemory<char>.Empty;
-            }
-
-            parser.Move();
-            var bparser = parser.BufferedView();
-            var skipResult = bparser.SkipUntil([')']);
-            if (skipResult.EndOfInput)
-            {
-                JustThrow("subgroup number");
-            }
-
-            var ret = parser.SourceUntilExclusive(bparser);
-            parser.MovePast(bparser.Position);
-
-            return ret;
-        }
-
-        [DoesNotReturn]
-        static void JustThrow(string part)
-        {
-            throw new NotSupportedException($"Bad {part}");
-        }
-    }
-
-    private struct SkipUntilOpenParenOrWhiteSpace : IShouldSkip
-    {
-        public bool ShouldSkip(char c)
-        {
-            if (c == '(')
-            {
-                return false;
-            }
-            if (char.IsWhiteSpace(c))
-            {
-                return false;
-            }
-            return true;
-        }
-    }
-
-    private static GroupId FindGroupMatch(Schedule schedule, in GroupForSearch g)
-    {
-        var groups = schedule.Groups;
-        for (int i = 0; i < groups.Length; i++)
-        {
-            var group = groups[i];
-            if (IsMatch(group, g))
-            {
-                return new(i);
-            }
-        }
-        return GroupId.Invalid;
-    }
-
-    private static bool IsMatch(Group a, in GroupForSearch b)
-    {
-        bool facultyMatches = a.Faculty.Name.AsSpan().Equals(
-            b.FacultyName.Span,
-            StringComparison.OrdinalIgnoreCase);
-        if (!facultyMatches)
-        {
-            return false;
-        }
-
-        if (a.GroupNumber != b.GroupNumber)
-        {
-            return false;
-        }
-
-        if (a.AttendanceMode != b.AttendanceMode)
-        {
-            return false;
-        }
-
-        if (a.QualificationType != b.QualificationType)
-        {
-            return false;
-        }
-
-        if (a.Grade != b.Grade)
-        {
-            return false;
-        }
-
-        return true;
-    }
-
-    private static IEnumerable<RegularLessonId> MatchLessons(LessonMatchParams p)
-    {
-        var lessonsOfCourse = p.Lookup[p.CourseId];
-        foreach (var lessonId in lessonsOfCourse)
-        {
-            var lesson = p.Schedule.Get(lessonId);
-            if (!lesson.Lesson.Groups.Contains(p.GroupId))
-            {
-                continue;
-            }
-            if (lesson.Lesson.SubGroup != p.SubGroup)
-            {
-                continue;
-            }
-
-            yield return lessonId;
-        }
-    }
-
-    // Intentionally duplicated, because the strings are actually different.
-    private static LessonType ParseLessonType(
-        string s,
-        ILogger logger)
-    {
-        var parser = new Parser(s);
-        parser.SkipWhitespace();
-        if (parser.IsEmpty)
-        {
-            return LessonType.Unspecified;
-        }
-        var bparser = parser.BufferedView();
-        _ = parser.SkipNotWhitespace();
-        var lessonTypeSpan = parser.PeekSpanUntilPosition(bparser.Position);
-        var lessonType = Get(lessonTypeSpan);
-        if (lessonType == LessonType.Custom)
-        {
-            logger.CustomLessonType(lessonTypeSpan);
-        }
-
-        parser.MoveTo(bparser.Position);
-
-        parser.SkipWhitespace();
-        if (!parser.IsEmpty)
-        {
-            throw new NotSupportedException("Lesson type not parsed fully.");
-        }
-        return lessonType;
-
-        LessonType Get(ReadOnlySpan<char> str)
-        {
-            static bool Equal(
-                ReadOnlySpan<char> str,
-                string literal)
-            {
-                return str.Equals(
-                    literal.AsSpan(),
-                    StringComparison.Ordinal);
-            }
-
-            if (Equal(str, "laborator"))
-            {
-                return LessonType.Lab;
-            }
-            if (Equal(str, "curs"))
-            {
-                return LessonType.Curs;
-            }
-            if (Equal(str, "seminar"))
-            {
-                return LessonType.Seminar;
-            }
-            return LessonType.Custom;
-        }
-    }
-
-    private readonly record struct LessonInstance
-    {
-        public required RegularLessonId LessonId { get; init; }
-        public required DateTime DateTime { get; init; }
-        // TODO: add topics
-    }
-
-    private readonly struct GetDateTimesOfScheduledLessonsParams
-    {
-        public required IEnumerable<RegularLessonId> Lessons { get; init; }
-        public required Schedule Schedule { get; init; }
-        public required LessonTimeConfig TimeConfig { get; init; }
-        public required IAllScheduledDateProvider DateProvider { get; init; }
-    }
-
-    private static IEnumerable<LessonInstance> GetDateTimesOfScheduledLessons(
-        GetDateTimesOfScheduledLessonsParams p)
-    {
-        foreach (var lessonId in p.Lessons)
-        {
-            var lesson = p.Schedule.Get(lessonId);
-            var lessonDate = lesson.Date;
-
-            var timeSlot = lessonDate.TimeSlot;
-            var startTime = p.TimeConfig.GetTimeSlotInterval(timeSlot).Start;
-
-            var dates = p.DateProvider.Dates(new()
-            {
-                Day = lessonDate.DayOfWeek,
-                Parity = lessonDate.Parity,
-            });
-            foreach (var date in dates)
-            {
-                var dateTime = new DateTime(
-                    date: date,
-                    time: startTime);
-                yield return new()
-                {
-                    LessonId = lessonId,
-                    DateTime = dateTime,
-                };
-            }
-        }
-    }
-
-    public struct ParseStudyWeekWordDocParams
-    {
-        public required string InputPath;
-    }
-
-    public static IEnumerable<StudyWeek> ParseStudyWeekWordDoc(ParseStudyWeekWordDocParams p)
+    public static ManualAllScheduledDateProvider CreateDateProviderFromWeekParityExcel(
+        ParseStudyWeekWordDocParams p)
     {
         using var stream = File.OpenRead(p.InputPath);
         using var word = WordprocessingDocument.Open(stream, isEditable: false);
-        if (word.MainDocumentPart?.Document is not { } document)
+        var studyWeeks = ParityExcelParser.Parse(word).ToArray();
+        var ret = new ManualAllScheduledDateProvider
         {
-            throw new InvalidOperationException("No document found.");
-        }
-        if (document.Body is not { } body)
-        {
-            throw new InvalidOperationException("No body found.");
-        }
-
-        var table = body.Descendants<Table>().First();
-        var rows = table.ChildElements.OfType<TableRow>();
-        using var rowEnumerator = rows.GetEnumerator();
-
-        DateOnly? previousWeekEnd = null;
-
-        VerifyHeader();
-        while (rowEnumerator.MoveNext())
-        {
-            var cells = GetDataCells(rowEnumerator.Current);
-            var weekInterval = Week();
-            var isOdd = IsOdd();
-            previousWeekEnd = weekInterval.End;
-            yield return new()
-            {
-                MondayDate = weekInterval.Start,
-                IsOddWeek = isOdd,
-            };
-            continue;
-
-            WeekInterval Week()
-            {
-                var weekText = cells.Week.InnerText;
-                var parser = new Parser(weekText);
-                var ret = ParseWeekInterval(ref parser);
-                parser.SkipWhitespace();
-                if (!parser.IsEmpty)
-                {
-                    throw new NotSupportedException("Invalid date range format.");
-                }
-
-                {
-                    int dayCount = ret.End.DayNumber - ret.Start.DayNumber + 1;
-                    if (dayCount != 6)
-                    {
-                        throw new NotSupportedException("The date range must have 6 days in total");
-                    }
-                }
-
-                if (previousWeekEnd is { } prevEnd)
-                {
-                    var dayDiff = ret.Start.DayNumber - prevEnd.DayNumber;
-                    if (dayDiff < 0)
-                    {
-                        throw new NotSupportedException("The week intervals must be in order.");
-                    }
-                }
-
-                if (ret.Start.DayOfWeek != DayOfWeek.Monday)
-                {
-                    throw new NotSupportedException("The week interval must start on a Monday.");
-                }
-
-                return ret;
-            }
-
-            bool IsOdd()
-            {
-                var parityText = cells.Parity.InnerText;
-                var parser = new Parser(parityText);
-                var ret = ParseIsOdd(ref parser);
-                parser.SkipWhitespace();
-                if (!parser.IsEmpty)
-                {
-                    throw new NotSupportedException("Invalid parity format.");
-                }
-                return ret;
-            }
-        }
-
-        yield break;
-
-
-        void VerifyHeader()
-        {
-            bool hasHeader = rowEnumerator.MoveNext();
-            if (!hasHeader)
-            {
-                throw new NotSupportedException("No header found.");
-            }
-
-            // Check has 3 columns
-            // Săptămâna
-            // ...
-            // Paritatea
-
-            var header = rowEnumerator.Current;
-            var headerCells = header.ChildElements.OfType<TableCell>();
-            using var cellEnumerator = headerCells.GetEnumerator();
-
-            bool CompareHeader(string expectedText)
-            {
-                var weekHeaderText = cellEnumerator.Current.InnerText.AsSpan().Trim();
-                return IgnoreDiacriticsComparer.Instance.Equals(weekHeaderText, expectedText.AsSpan());
-            }
-
-            {
-                bool moved = cellEnumerator.MoveNext();
-                if (!moved)
-                {
-                    throw new NotSupportedException("No row for the week string");
-                }
-
-                if (!CompareHeader("Saptamana"))
-                {
-                    throw new NotSupportedException("Week header should be first. It had unexpected name.");
-                }
-            }
-
-            {
-                bool moved = cellEnumerator.MoveNext();
-                if (!moved)
-                {
-                    throw new NotSupportedException("An insignificant column must follow after the week column.");
-                }
-            }
-
-            {
-                bool moved = cellEnumerator.MoveNext();
-                if (!moved)
-                {
-                    throw new NotSupportedException("No row for the parity string");
-                }
-
-                if (!CompareHeader("Paritatea"))
-                {
-                    throw new NotSupportedException("Week header should be first. It had unexpected name.");
-                }
-            }
-
-            {
-                bool moved = cellEnumerator.MoveNext();
-                if (moved)
-                {
-                    throw new NotSupportedException("Must only have 3 columns");
-                }
-            }
-        }
-
-        static (Cell Week, Cell Parity) GetDataCells(TableRow dataRow)
-        {
-            var cells = dataRow.ChildElements.OfType<Cell>();
-            using var cellsEnumerator = cells.GetEnumerator();
-
-            var weekCell = NextCell(cellsEnumerator);
-            _ = NextCell(cellsEnumerator);
-            var parityCell = NextCell(cellsEnumerator);
-            NoNextCell(cellsEnumerator);
-            return (
-                Week: weekCell,
-                Parity: parityCell);
-        }
-
-        static Cell NextCell(IEnumerator<Cell> cells)
-        {
-            bool moved = cells.MoveNext();
-            if (!moved)
-            {
-                throw new NotSupportedException("Expected a cell??");
-            }
-            return cells.Current;
-        }
-        static void NoNextCell(IEnumerator<Cell> cells)
-        {
-            bool moved = cells.MoveNext();
-            if (moved)
-            {
-                throw new NotSupportedException("Expected no more cells.");
-            }
-        }
-    }
-
-    private static readonly ImmutableArray<string> MonthNames = GetMonthNames();
-
-    private static ImmutableArray<string> GetMonthNames()
-    {
-        // ReSharper disable once CollectionNeverUpdated.Local
-        var ret = ImmutableArray.CreateBuilder<string>();
-        ret.Capacity = (int) Month.Count;
-        ret.Count = (int) Month.Count;
-        Set(Month.January, "ianuarie");
-        Set(Month.February, "februarie");
-        Set(Month.March, "martie");
-        Set(Month.April, "aprilie");
-        Set(Month.May, "mai");
-        Set(Month.June, "iunie");
-        Set(Month.July, "iulie");
-        Set(Month.August, "august");
-        Set(Month.September, "septembrie");
-        Set(Month.October, "octombrie");
-        Set(Month.November, "noiembrie");
-        Set(Month.December, "decembrie");
-        Debug.Assert(ret.All(x => x != null));
-        return ret.MoveToImmutable();
-
-        void Set(Month m, string name)
-        {
-            ret[(int) (m - 1)] = name;
-        }
-    }
-
-
-    private static Month? ParseMonth(ReadOnlySpan<char> name)
-    {
-        for (int i = 0; i < MonthNames.Length; i++)
-        {
-            var monthName = MonthNames[i];
-            if (monthName.AsSpan().Equals(name, StringComparison.OrdinalIgnoreCase))
-            {
-                var month = (Month)(i + 1);
-                return month;
-            }
-        }
-        return null;
-    }
-
-    private record struct WeekInterval(DateOnly Start, DateOnly End);
-
-    private readonly struct SkipPunctuationOrWhite : IShouldSkip
-    {
-        public bool ShouldSkip(char c)
-        {
-            if (char.IsPunctuation(c))
-            {
-                return true;
-            }
-            if (char.IsWhiteSpace(c))
-            {
-                return true;
-            }
-            return false;
-        }
-    }
-
-    private static WeekInterval ParseWeekInterval(ref Parser parser)
-    {
-        parser.SkipWhitespace();
-
-        var dayStart = ParseDayNumber(ref parser);
-        parser.SkipWhitespace();
-        var monthStart = ParseMonth1(ref parser);
-        parser.Skip(new SkipPunctuationOrWhite());
-
-        var dayEnd = ParseDayNumber(ref parser);
-        parser.SkipWhitespace();
-        var monthEnd = ParseMonth1(ref parser);
-        parser.SkipWhitespace();
-
-        var year = ParseYear(ref parser);
-        var startDate = CreateDate(dayStart, monthStart);
-        var endDate = CreateDate(dayEnd, monthEnd);
-        return new()
-        {
-            Start = startDate,
-            End = endDate,
+            StudyWeeks = studyWeeks,
         };
-
-        DateOnly CreateDate(uint day, Month month)
-        {
-            return new DateOnly(
-                year: year,
-                month: (int) month,
-                day: (int) day);
-        }
-
-        uint ParseDayNumber(ref Parser parser)
-        {
-            var bparser = parser.BufferedView();
-            var skipResult = bparser.SkipNumbers();
-            if (!skipResult.SkippedAny)
-            {
-                throw new NotSupportedException("Day number expected");
-            }
-            var daySpan = parser.PeekSpanUntilPosition(bparser.Position);
-            if (daySpan.Length > 2)
-            {
-                throw new NotSupportedException("Day number too long (max 2 numbers)");
-            }
-            if (!uint.TryParse(daySpan, out uint day))
-            {
-                Debug.Fail("This should never happen?");
-                day = 0;
-            }
-            parser.MoveTo(bparser.Position);
-            return day;
-        }
-
-        Month ParseMonth1(ref Parser parser)
-        {
-            var bparser = parser.BufferedView();
-            var skipResult = bparser.SkipLetters();
-            if (!skipResult.SkippedAny)
-            {
-                throw new NotSupportedException("Month name expected");
-            }
-            var monthSpan = parser.PeekSpanUntilPosition(bparser.Position);
-            if (ParseMonth(monthSpan) is not { } month)
-            {
-                throw new NotSupportedException("Month name not recognized");
-            }
-            return month;
-        }
-
-        int ParseYear(ref Parser parser)
-        {
-            var yearResult = parser.ConsumePositiveInt(4);
-            if (yearResult.Status != ConsumeIntStatus.Ok)
-            {
-                throw new NotSupportedException("Year not parsed");
-            }
-            return (int) yearResult.Value;
-        }
+        return ret;
     }
-
-    private static bool ParseIsOdd(ref Parser parser)
-    {
-        var bparser = parser.BufferedView();
-        bparser.SkipNotWhitespace();
-        var span = parser.PeekSpanUntilPosition(bparser.Position);
-
-        static bool Equals1(ReadOnlySpan<char> span, string literal)
-        {
-            return span.Equals(literal.AsSpan(), StringComparison.CurrentCultureIgnoreCase);
-        }
-
-        if (Equals1(span, "Pară"))
-        {
-            parser.MoveTo(bparser.Position);
-            return false;
-        }
-        if (Equals1(span, "Impară"))
-        {
-            parser.MoveTo(bparser.Position);
-            return true;
-        }
-        throw new NotSupportedException("Parity not recognized");
-    }
-}
-
-public enum Month
-{
-    January = 1,
-    February = 2,
-    March = 3,
-    April = 4,
-    May = 5,
-    June = 6,
-    July = 7,
-    August = 8,
-    September = 9,
-    October = 10,
-    November = 11,
-    December = 12,
-    Count = 12,
 }
 
 public enum Option
@@ -2157,143 +970,4 @@ public enum Option
     AllTeachersExcel,
     PerGroupAndPerTeacherPdfs,
     CreateLessonsInRegistry,
-}
-
-public sealed class Credentials
-{
-    public required string Login { get; set; }
-    public required string Password { get; set; }
-}
-
-public sealed class TokenCookieModel
-{
-    public required string Value { get; set; }
-    public DateTime ExpireTime { get; set; }
-
-    public Cookie ToObject(string name)
-    {
-        var ret = new Cookie(name, Value)
-        {
-            Expires = ExpireTime,
-        };
-        return ret;
-    }
-
-    public static TokenCookieModel FromObject(Cookie cookie)
-    {
-        var expireTime = cookie.Expires;
-        return new()
-        {
-            Value = cookie.Value,
-            ExpireTime = expireTime,
-        };
-    }
-}
-
-public enum Session
-{
-    Ses1,
-    Ses2,
-}
-
-internal readonly record struct CourseLink(CourseId CourseId, Uri Url);
-internal readonly record struct GroupLink
-{
-    public required GroupId GroupId { get; init; }
-    public required SubGroup SubGroup { get; init; }
-    public required Uri Uri { get; init; }
-}
-
-internal readonly record struct LessonInstanceLink
-{
-    public required DateTime DateTime { get; init; }
-    public required LessonType LessonType { get; init; }
-    public required Uri EditUri { get; init; }
-}
-
-internal readonly record struct LessonMatchParams
-{
-    public required CourseId CourseId { get; init; }
-    public required GroupId GroupId { get; init; }
-    public required SubGroup SubGroup { get; init; }
-    public required LessonsByCourseMap Lookup { get; init; }
-    public required Schedule Schedule { get; init; }
-}
-
-internal record struct LanguageOrFR
-{
-    public Language? Language;
-    public bool FR;
-    public readonly bool IsLanguage => Language is not null;
-}
-
-internal struct GroupForSearch
-{
-    public required AttendanceMode AttendanceMode;
-    public required Grade Grade;
-    public required int GroupNumber;
-    public required ReadOnlyMemory<char> FacultyName;
-    public required QualificationType QualificationType;
-    public required ReadOnlyMemory<char> SubGroupName;
-}
-
-public record struct StudyWeek(DateOnly MondayDate, bool IsOddWeek);
-
-
-public interface IAllScheduledDateProvider
-{
-
-    public readonly struct Params
-    {
-        public required Parity Parity { get; init; }
-        public required DayOfWeek Day { get; init; }
-    }
-    IEnumerable<DateOnly> Dates(Params p);
-}
-
-public sealed class ManualAllScheduledDateProvider
-    : IAllScheduledDateProvider
-{
-    public required StudyWeek[] StudyWeeks { private get; init; }
-
-    public IEnumerable<DateOnly> Dates(IAllScheduledDateProvider.Params p)
-    {
-        foreach (var week in StudyWeeks)
-        {
-            if (!IsParityMatch())
-            {
-                continue;
-            }
-            const int weekdayCount = 7;
-            var offset = (p.Day - DayOfWeek.Monday + weekdayCount) % weekdayCount;
-            var ret = week.MondayDate.AddDays(offset);
-            yield return ret;
-
-            continue;
-
-            bool IsParityMatch()
-            {
-                switch (p.Parity)
-                {
-                    case Parity.EveryWeek:
-                    {
-                        return true;
-                    }
-                    case Parity.OddWeek:
-                    {
-                        return week.IsOddWeek;
-                    }
-                    case Parity.EvenWeek:
-                    {
-                        return !week.IsOddWeek;
-                    }
-                    default:
-                    {
-                        Debug.Fail("Impossible value of parity");
-                        return false;
-                    }
-                }
-            }
-        }
-    }
 }
